+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - Added check detection in quiescence search
  - Added evasion moves generator
  - Added static null-move pruning
+ - Added futility pruning
 
 # Version 3.0 (Luna), 12.12.2020
  - Added better time control for Arbiter